feat: refactor, fixes, and documentations. - docs: OpenAPI 2.0 and code documentations. - api/orders: move from direct model binding to request binding. - models: move binding tags to its respective request. - order_actions/update: handle no deletion on update correctly. - order_routes: /prune -> /__prune__ - database: minor naming changes, use log.Fatal instead of panic

+++ api/orders/actions/update.go
@@ -6,12 +6,14 @@ import (
 	. "rockyprabowo/assignment-2/models"
 )
 
+// UpdateOrder updates an order on the database.
 func (actions OrderActions) UpdateOrder(order, updatePayload *Order) error {
 	originalOrderItemIDs := slices.Map(order.Items, ItemMapID)
 	order.CustomerName = updatePayload.CustomerName
 	order.Items = updatePayload.Items
 
-	err := actions.Database.
+	err := actions.
+		Database.
 		Session(&gorm.Session{FullSaveAssociations: true}).
 		Save(order).
 		Error
@@ -22,7 +24,9 @@ func (actions OrderActions) UpdateOrder(order, updatePayload *Order) error {
 	savedOrderItemIDs := slices.Map(order.Items, ItemMapID)
 
 	deletedItemIDs := slices.Diff(originalOrderItemIDs, savedOrderItemIDs)
+	if len(deletedItemIDs) > 0 {
+		actions.Database.Delete(&Item{}, deletedItemIDs)
+	}
 
-	actions.Database.Delete(&Item{}, deletedItemIDs)
 	return nil
 }
